WIP Simplify SideDrawer --> move stuff in subcomponents when possible
